feat(svg): extract parser
