+++ packages/image/src/index.ts
@@ -1,12 +1,5 @@
 import resolveImage from './resolve';
 
-export type {
-  Image,
-  ImageSrc,
-  SvgImage,
-  RasterImage,
-  SvgNode,
-  Viewbox,
-} from './types';
+export type { Image, ImageSrc, SvgImage, RasterImage } from './types';
 
 export default resolveImage;
