+++ packages/svg/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@react-pdf/image",
-  "version": "3.0.4",
+  "name": "@react-pdf/svg",
+  "version": "1.0.0",
   "license": "MIT",
-  "description": "Parses the images in png or jpeg format for react-pdf document",
+  "description": "SVG parsing for react-pdf",
   "author": "Diego Muracciole <diegomuracciole@gmail.com>",
   "homepage": "https://github.com/diegomura/react-pdf#readme",
   "type": "module",
@@ -14,18 +14,13 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/diegomura/react-pdf.git",
-    "directory": "packages/image"
+    "directory": "packages/svg"
   },
   "scripts": {
     "test": "vitest",
     "build": "rimraf ./lib && rollup -c",
     "watch": "rimraf ./lib && rollup -c -w",
     "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@react-pdf/png-js": "^3.0.0",
-    "@react-pdf/svg": "^1.0.0",
-    "jay-peg": "^1.1.1"
   },
   "files": [
     "lib"
