+++ packages/svg/rollup.config.js
@@ -1,0 +1,34 @@
+import { dts } from 'rollup-plugin-dts';
+import del from 'rollup-plugin-delete';
+import typescript from '@rollup/plugin-typescript';
+import replace from '@rollup/plugin-replace';
+
+const input = './src/index.ts';
+
+const getPlugins = ({ browser }) => [
+  typescript(),
+  replace({
+    preventAssignment: true,
+    values: { BROWSER: JSON.stringify(browser) },
+  }),
+];
+
+const serverConfig = {
+  input,
+  output: { format: 'es', file: 'lib/index.js' },
+  plugins: getPlugins({ browser: false }),
+};
+
+const browserConfig = {
+  input,
+  output: { format: 'es', file: 'lib/index.browser.js' },
+  plugins: getPlugins({ browser: true }),
+};
+
+const dtsConfig = {
+  input: './lib/types/index.d.ts',
+  output: [{ file: 'lib/index.d.ts', format: 'es' }],
+  plugins: [dts(), del({ targets: 'lib/types', hook: 'buildEnd' })],
+};
+
+export default [serverConfig, browserConfig, dtsConfig];
