+++ packages/svg/src/index.ts
@@ -1,0 +1,3 @@
+export { parseSvg } from './svg';
+export type { ParsedSvg } from './svg';
+export type { SvgNode, Viewbox } from './types';
