+++ packages/svg/src/types.ts
@@ -1,0 +1,12 @@
+export type Viewbox = {
+  minX: number;
+  minY: number;
+  maxX: number;
+  maxY: number;
+};
+
+export interface SvgNode {
+  type: string;
+  props: Record<string, unknown>;
+  children?: SvgNode[];
+}
