+++ packages/svg/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "https://json-schema.store.org/tsconfig",
+  "compilerOptions": {
+    "outDir": "lib",
+    "declaration": true,
+    "declarationDir": "lib/types",
+    "target": "ES2022",
+    "module": "ESNext",
+    "lib": ["ES2022", "DOM"],
+    "moduleResolution": "Node",
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "types": ["vitest/globals"],
+  },
+  "include": ["src", "**/*.d.ts"],
+}
