+++ packages/svg/vitest.config.js
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: ['tests/*.{test,spec}.?(c|m)[jt]s?(x)'],
+    watch: false,
+  },
+  define: {
+    BROWSER: false,
+  },
+});
